Adjust naming & add phpdocs
