+++ tests/NativeClassifierTest.php
@@ -4,13 +4,13 @@ declare(strict_types=1);
 
 namespace Yiisoft\Classifier\Tests;
 
+use Yiisoft\Classifier\NativeClassifier;
 use Yiisoft\Classifier\ClassifierInterface;
-use Yiisoft\Classifier\PhpParserClassifier;
 
-class PhpParserClassifierTest extends BaseClassifierTest
+class NativeClassifierTest extends BaseClassifierTest
 {
     protected function createClassifier(string ...$dirs): ClassifierInterface
     {
-        return new PhpParserClassifier(...$dirs);
+        return new NativeClassifier(...$dirs);
     }
 }
